pyproject.toml
[tool.poetry]
name = "floresce_py"
version = "0.1.0"
description = ""
authors = ["icedcoffeeee"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
pyinstaller = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
